LPD-56159 Move to model document contributor
