+++ modules/apps/object/object-service/src/main/java/com/liferay/object/internal/search/spi/model/index/contributor/ObjectEntryDocumentContributor.java
@@ -5,23 +5,14 @@
 
 package com.liferay.object.internal.search.spi.model.index.contributor;
 
-import com.liferay.object.constants.ObjectEntryFolderConstants;
 import com.liferay.object.model.ObjectDefinition;
 import com.liferay.object.model.ObjectEntry;
-import com.liferay.object.model.ObjectEntryFolder;
 import com.liferay.object.service.ObjectDefinitionLocalService;
-import com.liferay.object.service.ObjectEntryFolderLocalService;
 import com.liferay.object.service.ObjectEntryLocalService;
-import com.liferay.petra.string.CharPool;
-import com.liferay.portal.kernel.feature.flag.FeatureFlagManagerUtil;
 import com.liferay.portal.kernel.model.BaseModel;
 import com.liferay.portal.kernel.search.Document;
 import com.liferay.portal.kernel.search.DocumentContributor;
 import com.liferay.portal.kernel.search.Field;
-import com.liferay.portal.kernel.util.GetterUtil;
-import com.liferay.portal.kernel.util.StringUtil;
-
-import java.util.Objects;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -53,93 +44,10 @@ public class ObjectEntryDocumentContributor
 		field.setValue(objectDefinition.getClassName());
 
 		document.add(field);
-
-		if (FeatureFlagManagerUtil.isEnabled(
-				objectEntry.getCompanyId(), "LPD-17564")) {
-
-			_contributeObjectEntryFolderAttributes(
-				document, objectEntry.getObjectEntryFolderId());
-		}
-	}
-
-	private void _contributeObjectEntryFolderAttributes(
-		Document document, long objectEntryFolderId) {
-
-		document.addKeyword(Field.FOLDER_ID, objectEntryFolderId);
-
-		ObjectEntryFolder objectEntryFolder =
-			_objectEntryFolderLocalService.fetchObjectEntryFolder(
-				objectEntryFolderId);
-
-		if (objectEntryFolder == null) {
-			document.addKeyword("cms_section", "none");
-		}
-		else {
-			document.addKeyword("cms_kind", "object");
-
-			ObjectEntryFolder rootObjectEntryFolder = _getRootObjectEntryFolder(
-				objectEntryFolder);
-
-			document.addKeyword(
-				"cms_root",
-				rootObjectEntryFolder.getObjectEntryFolderId() ==
-					objectEntryFolderId);
-			document.addKeyword(
-				"cms_section",
-				_getCMSSection(
-					rootObjectEntryFolder.getExternalReferenceCode()));
-		}
-	}
-
-	private String _getCMSSection(String externalReferenceCode) {
-		if (externalReferenceCode.equals(
-				ObjectEntryFolderConstants.EXTERNAL_REFERENCE_CODE_CONTENTS)) {
-
-			return "contents";
-		}
-
-		if (externalReferenceCode.equals(
-				ObjectEntryFolderConstants.EXTERNAL_REFERENCE_CODE_FILES)) {
-
-			return "files";
-		}
-
-		return "none";
-	}
-
-	private ObjectEntryFolder _getRootObjectEntryFolder(
-		ObjectEntryFolder objectEntryFolder) {
-
-		if (objectEntryFolder == null) {
-			return null;
-		}
-
-		if (Objects.equals(
-				objectEntryFolder.getExternalReferenceCode(),
-				ObjectEntryFolderConstants.EXTERNAL_REFERENCE_CODE_CONTENTS) ||
-			Objects.equals(
-				objectEntryFolder.getExternalReferenceCode(),
-				ObjectEntryFolderConstants.EXTERNAL_REFERENCE_CODE_FILES)) {
-
-			return objectEntryFolder;
-		}
-
-		String[] parts = StringUtil.split(
-			objectEntryFolder.getTreePath(), CharPool.SLASH);
-
-		if (parts.length <= 2) {
-			return null;
-		}
-
-		return _objectEntryFolderLocalService.fetchObjectEntryFolder(
-			GetterUtil.getLong(parts[1]));
 	}
 
 	@Reference
 	private ObjectDefinitionLocalService _objectDefinitionLocalService;
-
-	@Reference
-	private ObjectEntryFolderLocalService _objectEntryFolderLocalService;
 
 	@Reference
 	private ObjectEntryLocalService _objectEntryLocalService;
